Reorganised codebase and added send and wait
