+++ src/server/tools.ts
@@ -1,0 +1,80 @@
+import { Tool } from '@modelcontextprotocol/sdk/types.js';
+
+export const tools: Tool[] = [
+  {
+    name: "send-message",
+    description: "Send a message to a Discord channel",
+    inputSchema: {
+      type: "object",
+      properties: {
+        server: {
+          type: "string",
+          description: 'Server name or ID (optional if bot is only in one server)',
+        },
+        channel: {
+          type: "string",
+          description: 'Channel name (e.g., "general") or ID',
+        },
+        message: {
+          type: "string",
+          description: "Message content to send",
+        },
+      },
+      required: ["channel", "message"],
+    },
+  },
+  {
+    name: "send-and-wait",
+    description: "Send a message to a Discord channel and wait for a response from a specific user",
+    inputSchema: {
+      type: "object",
+      properties: {
+        server: {
+          type: "string",
+          description: 'Server name or ID (optional if bot is only in one server)',
+        },
+        channel: {
+          type: "string",
+          description: 'Channel name (e.g., "general") or ID',
+        },
+        message: {
+          type: "string",
+          description: "Message content to send",
+        },
+        userId: {
+          type: "string",
+          description: "Discord user ID to wait for response from",
+        },
+        timeout: {
+          type: "number",
+          description: "Timeout in milliseconds to wait for response (default: 60000)",
+          default: 60000,
+        },
+      },
+      required: ["channel", "message", "userId"],
+    },
+  },
+  {
+    name: "read-messages",
+    description: "Read recent messages from a Discord channel",
+    inputSchema: {
+      type: "object",
+      properties: {
+        server: {
+          type: "string",
+          description: 'Server name or ID (optional if bot is only in one server)',
+        },
+        channel: {
+          type: "string",
+          description: 'Channel name (e.g., "general") or ID',
+        },
+        limit: {
+          type: "number",
+          description: "Number of messages to fetch (max 100)",
+          default: 50,
+        },
+      },
+      required: ["channel"],
+    },
+  },
+];
