+++ src/utils/discord.ts
@@ -1,0 +1,106 @@
+import { Client, Guild, TextChannel, Message } from 'discord.js';
+import { MessageResponse } from '../types/schemas.js';
+
+export async function findGuild(client: Client, guildIdentifier?: string): Promise<Guild> {
+  if (!guildIdentifier) {
+    // If no guild specified and bot is only in one guild, use that
+    if (client.guilds.cache.size === 1) {
+      return client.guilds.cache.first()!;
+    }
+    // List available guilds
+    const guildList = Array.from(client.guilds.cache.values())
+      .map(g => `"${g.name}"`).join(', ');
+    throw new Error(`Bot is in multiple servers. Please specify server name or ID. Available servers: ${guildList}`);
+  }
+
+  // Try to fetch by ID first
+  try {
+    const guild = await client.guilds.fetch(guildIdentifier);
+    if (guild) return guild;
+  } catch {
+    // If ID fetch fails, search by name
+    const guilds = client.guilds.cache.filter(
+      g => g.name.toLowerCase() === guildIdentifier.toLowerCase()
+    );
+    
+    if (guilds.size === 0) {
+      const availableGuilds = Array.from(client.guilds.cache.values())
+        .map(g => `"${g.name}"`).join(', ');
+      throw new Error(`Server "${guildIdentifier}" not found. Available servers: ${availableGuilds}`);
+    }
+    if (guilds.size > 1) {
+      const guildList = guilds.map(g => `${g.name} (ID: ${g.id})`).join(', ');
+      throw new Error(`Multiple servers found with name "${guildIdentifier}": ${guildList}. Please specify the server ID.`);
+    }
+    return guilds.first()!;
+  }
+  throw new Error(`Server "${guildIdentifier}" not found`);
+}
+
+export async function findChannel(client: Client, channelIdentifier: string, guildIdentifier?: string): Promise<TextChannel> {
+  const guild = await findGuild(client, guildIdentifier);
+  
+  // First try to fetch by ID
+  try {
+    const channel = await client.channels.fetch(channelIdentifier);
+    if (channel instanceof TextChannel && channel.guild.id === guild.id) {
+      return channel;
+    }
+  } catch {
+    // If fetching by ID fails, search by name in the specified guild
+    const channels = guild.channels.cache.filter(
+      (channel): channel is TextChannel =>
+        channel instanceof TextChannel &&
+        (channel.name.toLowerCase() === channelIdentifier.toLowerCase() ||
+         channel.name.toLowerCase() === channelIdentifier.toLowerCase().replace('#', ''))
+    );
+
+    if (channels.size === 0) {
+      const availableChannels = guild.channels.cache
+        .filter((c): c is TextChannel => c instanceof TextChannel)
+        .map(c => `"#${c.name}"`).join(', ');
+      throw new Error(`Channel "${channelIdentifier}" not found in server "${guild.name}". Available channels: ${availableChannels}`);
+    }
+    if (channels.size > 1) {
+      const channelList = channels.map(c => `#${c.name} (${c.id})`).join(', ');
+      throw new Error(`Multiple channels found with name "${channelIdentifier}" in server "${guild.name}": ${channelList}. Please specify the channel ID.`);
+    }
+    return channels.first()!;
+  }
+  throw new Error(`Channel "${channelIdentifier}" is not a text channel or not found in server "${guild.name}"`);
+}
+
+export function waitForResponse(
+  client: Client,
+  channel: TextChannel,
+  userId: string,
+  timeout: number = 60000
+): Promise<MessageResponse> {
+  return new Promise((resolve, reject) => {
+    const timeoutId = setTimeout(() => {
+      cleanup();
+      reject(new Error(`Timeout waiting for response from user ${userId}`));
+    }, timeout);
+
+    const messageHandler = (message: Message) => {
+      if (message.channelId === channel.id && message.author.id === userId) {
+        cleanup();
+        resolve({
+          content: message.content,
+          author: {
+            id: message.author.id,
+            tag: message.author.tag,
+          },
+          timestamp: message.createdAt.toISOString(),
+        });
+      }
+    };
+
+    const cleanup = () => {
+      clearTimeout(timeoutId);
+      client.off('messageCreate', messageHandler);
+    };
+
+    client.on('messageCreate', messageHandler);
+  });
+}
